Copied ctack.cmd from desktop
The main test script to be examined.  It follows the instructions given to me through email.  For the unzipping, copying, and deleting, the operations were done on example files inside a folder called C:\Example.  So this code could be modified accordingly depending on what file a user wanted to unzip or where to copy the resulting directory, and so forth.  The timing data results are also printed in a text file inside this C:\Example folder.

More info is included in the README.